Introduce the notion of the "main" module, i.e. the module which defines main().  Disallow top-level statements in modules other than the main module.  Fix IRGen to generate a main() function in the main module, and implement top-level statements.  Get rid of the main() mangling hack.  Part of <rdar://problem/11185451>.
I haven't really carefully considered whether existing type-checking etc. is correct for the main module (I think the name-binding rules need to be a bit different?), but it seems to work well enough for the obvious cases.

This is enough to get the one-liner "println(10)" to print "10" in "swift -i" mode, although the path to swift.swift still needs to be explicitly provided with -I.



Swift SVN r1325