WIP: Part one of ObjC briding
swift_StringToNSString() should be done and swift_NSStringToString()
should work with a "_NSSwiftString" produced by the former.

Work still needs to be done on extracting/bridging other NSStrings.

There are probably bugs, in particular with which side of the bridge
does retain/release calls on the boxed Swift String.

Also, yes, this is mess and will need to be cleaned up.

Swift SVN r1929