First draft of some code to make the general unqualified lookup able to perform lookup into local scopes.  Normally, name-binding doesn't need this because the parser can resolve references to locals, but we need this sort of lookup to handle cases involving local types.  It's also likely this will be generally useful for tools consuming the AST, typo-correction, etc.
We probably need to add some sort of data structure to represent this information, but as a proof of concept the current code appears to work.  I'm still working out how to make sure the parser doesn't prematurely bind names and how to make name-binding use it where appropriate (and avoid it when we don't need it, because no matter how efficient we make it, it will still be relatively expensive).



Swift SVN r2112