Eliminate SuperConversionExpr. This expression node was only really used in the very narrow case where we were converting from one protocol type to another (super) protocol type. However, ErasureExpr now handles this case via its null conformance entries (for the "trivial" cases), and can cope with general existential types where some conversions are trivial and others are not.
The IR generation side of this is basically just a hack to inline the
existing super-conversion code into the erasure code. This whole
routine will eventually need to be reworked anyway to deal with
destination types that are protocol-conformance types and with source
types that are archetypes (for generic/existential interactions).



Swift SVN r2213