Teach the overloaded declaration reference expression to track whether the base of the expression is ignored, as it will be for references to static member functions or when the base is actually of metatype type. Then, always also one to access the type of the base, even if it's a metatype type.
Use this information to simplify our handling of archetype member
reference expressions, which can refer to either members of the
meta-archetype or of an instance of the archetype. This allows us to
refer to an instance methodr of an archetype type (e.g., T.method) and
get back its curried form ([byref] T) -> inputs -> result.



Swift SVN r2218