Fix a conceptual bug in class metadata:  the parent pointer of a class is part of the class members section and is not global to the entire class metadata.  This is crucial for correct operation of functions expecting a base-class metadata object.
That gives us the correct foundation to implement an
optimization under which generic arguments that can be
inferred from the 'this' pointer need not actually be
separately passed.  This has the important result of
making all class member functions with the same signature
up to abstraction actually have the same physical
signature.

Swift SVN r2936