Fix CMake build of httpd.
This involves a little hack: right now every file in a swift_executable
is compiled without -parse-as-library, but if multiple files have top-level
code, presumably we'll get linker errors. Ideally we'd have some way to
figure out which file is the main file without having to specify it
explicitly.

The other change is to mark Swift object files as "C", so that they get
linked like C object files.

Swift SVN r3036