Keep type variable bounds on-line rather than computing them later.
Switch from computing bounds on each type variable while we're solving
a constraint system to keeping these bounds on-line, so that we can
produce definitive bindings during simplification. At present, this is
actually a serious pessimization (25% slower on
swift.swift!). However, it should allow us to direct the constraint
solver better, e.g., by allowing us to solve all of the constraints
directly pertaining to a type variable when we have a guess for that
type variable.

If it doesn't pan out soon enough, I can easily revert it.




Swift SVN r3109