When definitively binding a type variable, visit its related constraints first.
This approach gives the solver just a little more direction,
re-simplifying constraints involving the just-bound type variable
first when we're definitively binding it. Note that we're only
handling this in the definitive binding case, and not the speculative
case, so we've only tackled the easier (and less rewarding) part of
the problem. Reduces the number of simplify() iterations by ~500
(3.7%) and Debug performance on swift.swift by < 1%.



Swift SVN r3114