Build child constraint systems when needed, rather than eagerly.
This has only a tiny impact on memory footprint (improving it
slightly) and no impact on performance. However, it makes the stack of
constraint systems act more like a stack, so that we don't do a little
computation on a constraint system and then revisit it later for the
rest of the computation.


Swift SVN r3150