Document the [objc], [IBOutlet], and [IBAction] attributes.
IBOutlet and IBAction are used by Xcode to support XIB editing, but moreover
there is not an /inherent/ requirement that they imply [objc]. For [IBAction],
it seems unlikely that this will ever change, but [IBOutlet] already tries
several different ways to make its connection, and moreover with outlets for
which we can see the declaration, *the type of the object is known,*, and IB
could in theory know to use the Swift entry points.

IBOutlet and IBAction are currently recognized; my plan is to make them both
alias [objc] internally for now.

Swift SVN r3317