Eliminate my ill-fated renaming of parameters in performSelector:withObject:withObject:.
Adding numbers to the names of the parameters in the interface doesn't
maintain the selector mapping, and we don't have/want any other way to
provide the selector, so this whole idea was bogus. 

Instead, we now allow the re-use of parameter names in the interface,
and have taught the tuple-conversion code to allow matching
subsequence elements with the same name, e.g.,

  x.performSelector(sel, withObject=foo1, withObject=foo2)

TODO: Make sure that selector-style declarations allow this, so long
as the parameter names used in the body are distinct.


Swift SVN r3706