Mangle the <context> of an ObjC class as 'So'.
Previously, we were mangling [objc] class types as swift
classes (using the defining component in the mangling) and
imported class types as class types with empty contexts.
This could create an inconsistency between components which
disagree about whether an ObjC class is defined in swift or
not;  this will not do.  Also, an empty context creates an
invalid mangling.

Swift SVN r3812