Function declarations must have bodies
Back when we started the project, we didn't have modules, a standard
library, or even the "asmname" attribute. Not requring function bodies
provided a quick way to prototype the language. We're long past that day
now.

If you still want or need a pure declaration, the [asmname="..."]
attribute still does not require a function body.

This fixes:
13036833 Syntax ambiguity between selector function decl syntax and top-level code

Swift SVN r3821