Don't use _objc_empty_vtable in JIT mode.
In 10.9, the method cache now takes up two words, and the vtable is gone.
Code compiled statically will work fine, but runtime classes need to fill
the rest of the method cache with zero, instead of an address.

I tested this on 10.8 as well and it seems to do no harm.

Swift SVN r3910